+++ tinted-builder-rust/src/utils.rs
@@ -4,7 +4,7 @@ use std::io::Write;
 use std::path::Path;
 
 #[allow(dead_code)]
-pub fn write_to_file(path: &Path, contents: &str) -> Result<()> {
+pub(crate) fn write_to_file(path: &Path, contents: &str) -> Result<()> {
     if path.exists() {
         remove_file(path).with_context(|| format!("Unable to remove file: {}", path.display()))?;
     }